requirements.txt
apptools==5.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
colorama==0.4.4
configobj==5.0.6
Cython==0.29.23
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
envisage==6.0.0
euclid3==0.1
importlib-metadata==4.4.0
importlib-resources==5.1.4
ipydatawidgets==4.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
K3D==2.9.6
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mayavi==4.7.3
meshio==4.4.5
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.53.1
numpy==1.20.3
numpy-stl==2.16.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
prettytable==0.7.2
prometheus-client==0.11.0
prompt-toolkit==3.0.18
py-vox-io==0.1
pycparser==2.20
pyface==7.3.0
Pygments==2.9.0
PyMCubes==0.1.2
pynrrd==0.4.2
PyOpenGL==3.1.5
pyparsing==2.4.7
pypng==0.0.19
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
pywin32==301
pywinpty==1.1.1
pyzmq==22.1.0
quad-mesh-simplify==1.1.4
regex==2019.12.20
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
solidpython==1.1.1
terminado==0.10.0
testpath==0.5.0
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
traits==6.2.0
traitsui==7.2.1
traittypes==0.2.1
voxelfuse==1.2.7
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
